sumatoria2_q2n

+++ Guía4/sumatoriaq^i.hs
@@ -1,3 +1,3 @@
-sumatoriaqn :: Integer-> Integer -> Integer
+sumatoriaqn :: Float -> Integer -> Float
 sumatoriaqn q n | n == 1 = q 
                 | otherwise = sumatoriaqn q (n-1) + q^n
